Handle UnicodeDecodeError in check.
This file causes chardet to give a high confidence for an
encoding which python cannot decode the file as, which triggers
a UnicodeDecodeError. This should just result in is_binary returning
false but instead results in a KeyError.

The reason for this is that the error is thrown before the local
was assigned, and thus when it comes to formatting the string for
logging the local is not present and so you get the KeyError.