Add very simple Hypothesis based tests
This is essentially a "getting started" test. All it does is assert
that the is_binary function should never crash regardless of what's
in the file.

There should be more interesting things to test here, but most of
the additional things I've tried fail and it's unclear that these
are really bugs so much as edge cases in the heuristic.

Note: This is the test that caught ssue #12.