Handle LookupError when detecting encoding
For reasons that are a slight mystery to me, chardet can return
encodings that Python doesn't actually support.

On the assumption that if we can't decode it we're probably not
interested in considering it as text (especially given that chardet
will sometimes detect an encoding that you can't actually decode as
when its confidence is < 1), this handles this commit by
considering it to not be decodable.